==> 5_Transmission_DemStoch_Sim-AUC.R <==
#### DOC STRING ===============================================================
# This script (1) runs demographically stochastic single-parasite model 
# simulations and (2) uses time series outputs from the demographically 
# stochastic two-parasite model simulations to calculate total transmission of 
# a single parasite/per parasite for each simulation (1000). 

# If you choose to run this script, outputs will be stored in 
# "EnvironmentalVariability_Repo/Generated_Outputs".

# To see the outputs generated by this script without running it, go to 
# "EnvironmentalVariability_Repo/Outputs/TotalTransmission_Outputs".
#### END DOC STRING ===========================================================

# Required packages
install.packages("MESS"); library(MESS)

#### Single parasite ####
#### Low-average temperature environment ####
#### ... Identical TPCs ####
DiscreteModel_Dem <- function(Pi_Last, Pj_Last, I_Last){
  ts = 0.1
  ri = 2.13; rj = 0; k = 0.01; p = 1; o = 1000
  
  Pi_B = rpois(1, ri*Pi_Last*ts) 
  Pi_D = rbinom(1, Pi_Last, (1-exp(-k*I_Last*ts))) 
  Pj_B = rpois(1, rj*Pj_Last*ts) 
  Pj_D = rbinom(1, Pj_Last, (1-exp(-k*I_Last*ts)))
  I_B = rpois(1, p*((Pi_Last/(Pi_Last + o)) + (Pj_Last/(Pj_Last + o)))*I_Last*ts) 
  
  Pi_Next = Pi_Last + Pi_B - Pi_D
  Pj_Next = Pj_Last + Pj_B - Pj_D
  I_Next = I_Last + I_B
  
  Pi_Last <- Pi_Next
  Pj_Last <- Pj_Next
  I_Last <- I_Next
  
  ## To prevent the populations from going negative
  if (Pi_Last < 0){
    Pi_Last <- 0
  }
  if (Pj_Last < 0){
    Pj_Last <- 0
  }
  ## To get host death if parasite abundance exceeds D threshold set by Antia
  if (Pi_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  if (Pj_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  
  return(c(Pi_Last, Pj_Last, I_Last))
}

DemStoch <- data.frame(time = seq(0,50,0.1))
SimNum <- seq(1,1000,1)
TT <- 50/0.1
for (i in 1:length(SimNum)){
  df <- data.frame()
  Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
  for (j in 1:TT){
    Output = DiscreteModel_Dem(Pi_Last, Pj_Last, I_Last)
    Pi_Last = Output[1]
    Pj_Last = Output[2]
    I_Last = Output[3]
    Addition <- c(Pi_Last, Pj_Last, I_Last)
    df <- rbind(df, Addition)
    colnames(df) <- c(paste0("Pi_", SimNum[i]),
                      paste0("Pj_", SimNum[i]),
                      paste0("I_", SimNum[i]))
  }
  inits <- data.frame(1, 0, 1)
  colnames(inits) <- c(paste0("Pi_", SimNum[i]),
                       paste0("Pj_", SimNum[i]),
                       paste0("I_", SimNum[i]))
  df <- data.frame(rbind(inits, df))
  DemStoch <- data.frame(cbind(DemStoch, df))
}
saveRDS(DemStoch, file = "Generated_Outputs/D_LT_Iden_1P.rds")

#### ... Semi-Identical TPCs ####
DiscreteModel_Dem <- function(Pi_Last, Pj_Last, I_Last){
  ts = 0.1
  ri = 1.795; rj = 0; k = 0.01; p = 1; o = 1000
  
  Pi_B = rpois(1, ri*Pi_Last*ts) 
  Pi_D = rbinom(1, Pi_Last, (1-exp(-k*I_Last*ts))) 
  Pj_B = rpois(1, rj*Pj_Last*ts) 
  Pj_D = rbinom(1, Pj_Last, (1-exp(-k*I_Last*ts)))
  I_B = rpois(1, p*((Pi_Last/(Pi_Last + o)) + (Pj_Last/(Pj_Last + o)))*I_Last*ts) 
  
  Pi_Next = Pi_Last + Pi_B - Pi_D
  Pj_Next = Pj_Last + Pj_B - Pj_D
  I_Next = I_Last + I_B
  
  Pi_Last <- Pi_Next
  Pj_Last <- Pj_Next
  I_Last <- I_Next
  
  ## To prevent the populations from going negative
  if (Pi_Last < 0){
    Pi_Last <- 0
  }
  if (Pj_Last < 0){
    Pj_Last <- 0
  }
  ## To get host death if parasite abundance exceeds D threshold set by Antia
  if (Pi_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  if (Pj_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  
  return(c(Pi_Last, Pj_Last, I_Last))
}

DemStoch <- data.frame(time = seq(0,50,0.1))
SimNum <- seq(1,1000,1)
TT <- 50/0.1
for (i in 1:length(SimNum)){
  df <- data.frame()
  Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
  for (j in 1:TT){
    Output = DiscreteModel_Dem(Pi_Last, Pj_Last, I_Last)
    Pi_Last = Output[1]
    Pj_Last = Output[2]
    I_Last = Output[3]
    Addition <- c(Pi_Last, Pj_Last, I_Last)
    df <- rbind(df, Addition)
    colnames(df) <- c(paste0("Pi_", SimNum[i]),
                      paste0("Pj_", SimNum[i]),
                      paste0("I_", SimNum[i]))
  }
  inits <- data.frame(1, 0, 1)
  colnames(inits) <- c(paste0("Pi_", SimNum[i]),
                       paste0("Pj_", SimNum[i]),
                       paste0("I_", SimNum[i]))
  df <- data.frame(rbind(inits, df))
  DemStoch <- data.frame(cbind(DemStoch, df))
}
saveRDS(DemStoch, file = "Generated_Outputs/D_LT_Semi_1P.rds")

#### ... Pi-Advantage TPCs ####
DiscreteModel_Dem <- function(Pi_Last, Pj_Last, I_Last){
  ts = 0.1
  ri = 1.83; rj = 0; k = 0.01; p = 1; o = 1000
  
  Pi_B = rpois(1, ri*Pi_Last*ts) 
  Pi_D = rbinom(1, Pi_Last, (1-exp(-k*I_Last*ts))) 
  Pj_B = rpois(1, rj*Pj_Last*ts) 
  Pj_D = rbinom(1, Pj_Last, (1-exp(-k*I_Last*ts)))
  I_B = rpois(1, p*((Pi_Last/(Pi_Last + o)) + (Pj_Last/(Pj_Last + o)))*I_Last*ts) 
  
  Pi_Next = Pi_Last + Pi_B - Pi_D
  Pj_Next = Pj_Last + Pj_B - Pj_D
  I_Next = I_Last + I_B
  
  Pi_Last <- Pi_Next
  Pj_Last <- Pj_Next
  I_Last <- I_Next
  
  ## To prevent the populations from going negative
  if (Pi_Last < 0){
    Pi_Last <- 0
  }
  if (Pj_Last < 0){
    Pj_Last <- 0
  }
  ## To get host death if parasite abundance exceeds D threshold set by Antia
  if (Pi_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  if (Pj_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  
  return(c(Pi_Last, Pj_Last, I_Last))
}

DemStoch <- data.frame(time = seq(0,50,0.1))
SimNum <- seq(1,1000,1)
TT <- 50/0.1
for (i in 1:length(SimNum)){
  df <- data.frame()
  Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
  for (j in 1:TT){
    Output = DiscreteModel_Dem(Pi_Last, Pj_Last, I_Last)
    Pi_Last = Output[1]
    Pj_Last = Output[2]
    I_Last = Output[3]
    Addition <- c(Pi_Last, Pj_Last, I_Last)
    df <- rbind(df, Addition)
    colnames(df) <- c(paste0("Pi_", SimNum[i]),
                      paste0("Pj_", SimNum[i]),
                      paste0("I_", SimNum[i]))
  }
  inits <- data.frame(1, 0, 1)
  colnames(inits) <- c(paste0("Pi_", SimNum[i]),
                       paste0("Pj_", SimNum[i]),
                       paste0("I_", SimNum[i]))
  df <- data.frame(rbind(inits, df))
  DemStoch <- data.frame(cbind(DemStoch, df))
}
saveRDS(DemStoch, file = "Generated_Outputs/D_LT_PiAd_1P.rds")

#### ... Pj-Advantage, Verison A TPCs ####
DiscreteModel_Dem <- function(Pi_Last, Pj_Last, I_Last){
  ts = 0.1
  ri = 1.65; rj = 0; k = 0.01; p = 1; o = 1000
  
  Pi_B = rpois(1, ri*Pi_Last*ts) 
  Pi_D = rbinom(1, Pi_Last, (1-exp(-k*I_Last*ts))) 
  Pj_B = rpois(1, rj*Pj_Last*ts) 
  Pj_D = rbinom(1, Pj_Last, (1-exp(-k*I_Last*ts)))
  I_B = rpois(1, p*((Pi_Last/(Pi_Last + o)) + (Pj_Last/(Pj_Last + o)))*I_Last*ts) 
  
  Pi_Next = Pi_Last + Pi_B - Pi_D
  Pj_Next = Pj_Last + Pj_B - Pj_D
  I_Next = I_Last + I_B
  
  Pi_Last <- Pi_Next
  Pj_Last <- Pj_Next
  I_Last <- I_Next
  
  ## To prevent the populations from going negative
  if (Pi_Last < 0){
    Pi_Last <- 0
  }
  if (Pj_Last < 0){
    Pj_Last <- 0
  }
  ## To get host death if parasite abundance exceeds D threshold set by Antia
  if (Pi_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  if (Pj_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  
  return(c(Pi_Last, Pj_Last, I_Last))
}

DemStoch <- data.frame(time = seq(0,50,0.1))
SimNum <- seq(1,1000,1)
TT <- 50/0.1
for (i in 1:length(SimNum)){
  df <- data.frame()
  Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
  for (j in 1:TT){
    Output = DiscreteModel_Dem(Pi_Last, Pj_Last, I_Last)
    Pi_Last = Output[1]
    Pj_Last = Output[2]
    I_Last = Output[3]
    Addition <- c(Pi_Last, Pj_Last, I_Last)
    df <- rbind(df, Addition)
    colnames(df) <- c(paste0("Pi_", SimNum[i]),
                      paste0("Pj_", SimNum[i]),
                      paste0("I_", SimNum[i]))
  }
  inits <- data.frame(1, 0, 1)
  colnames(inits) <- c(paste0("Pi_", SimNum[i]),
                       paste0("Pj_", SimNum[i]),
                       paste0("I_", SimNum[i]))
  df <- data.frame(rbind(inits, df))
  DemStoch <- data.frame(cbind(DemStoch, df))
}
saveRDS(DemStoch, file = "Generated_Outputs/D_LT_PjAd_1P.rds")

#### High-average temperature environment ####
#### ... Identical TPCs ####
DiscreteModel_Dem <- function(Pi_Last, Pj_Last, I_Last){
  ts = 0.1
  ri = 2.15; rj = 0; k = 0.01; p = 1; o = 1000
  
  Pi_B = rpois(1, ri*Pi_Last*ts) 
  Pi_D = rbinom(1, Pi_Last, (1-exp(-k*I_Last*ts))) 
  Pj_B = rpois(1, rj*Pj_Last*ts) 
  Pj_D = rbinom(1, Pj_Last, (1-exp(-k*I_Last*ts)))
  I_B = rpois(1, p*((Pi_Last/(Pi_Last + o)) + (Pj_Last/(Pj_Last + o)))*I_Last*ts) 
  
  Pi_Next = Pi_Last + Pi_B - Pi_D
  Pj_Next = Pj_Last + Pj_B - Pj_D
  I_Next = I_Last + I_B
  
  Pi_Last <- Pi_Next
  Pj_Last <- Pj_Next
  I_Last <- I_Next
  
  ## To prevent the populations from going negative
  if (Pi_Last < 0){
    Pi_Last <- 0
  }
  if (Pj_Last < 0){
    Pj_Last <- 0
  }
  ## To get host death if parasite abundance exceeds D threshold set by Antia
  if (Pi_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  if (Pj_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  
  return(c(Pi_Last, Pj_Last, I_Last))
}

DemStoch <- data.frame(time = seq(0,50,0.1))
SimNum <- seq(1,1000,1)
TT <- 50/0.1
for (i in 1:length(SimNum)){
  df <- data.frame()
  Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
  for (j in 1:TT){
    Output = DiscreteModel_Dem(Pi_Last, Pj_Last, I_Last)
    Pi_Last = Output[1]
    Pj_Last = Output[2]
    I_Last = Output[3]
    Addition <- c(Pi_Last, Pj_Last, I_Last)
    df <- rbind(df, Addition)
    colnames(df) <- c(paste0("Pi_", SimNum[i]),
                      paste0("Pj_", SimNum[i]),
                      paste0("I_", SimNum[i]))
  }
  inits <- data.frame(1, 0, 1)
  colnames(inits) <- c(paste0("Pi_", SimNum[i]),
                       paste0("Pj_", SimNum[i]),
                       paste0("I_", SimNum[i]))
  df <- data.frame(rbind(inits, df))
  DemStoch <- data.frame(cbind(DemStoch, df))
}
saveRDS(DemStoch, file = "Generated_Outputs/D_HT_Iden_1P.rds")

#### ... Semi-Identical TPCs ####
DiscreteModel_Dem <- function(Pi_Last, Pj_Last, I_Last){
  ts = 0.1
  ri = 1.83; rj = 0; k = 0.01; p = 1; o = 1000
  
  Pi_B = rpois(1, ri*Pi_Last*ts) 
  Pi_D = rbinom(1, Pi_Last, (1-exp(-k*I_Last*ts))) 
  Pj_B = rpois(1, rj*Pj_Last*ts) 
  Pj_D = rbinom(1, Pj_Last, (1-exp(-k*I_Last*ts)))
  I_B = rpois(1, p*((Pi_Last/(Pi_Last + o)) + (Pj_Last/(Pj_Last + o)))*I_Last*ts) 
  
  Pi_Next = Pi_Last + Pi_B - Pi_D
  Pj_Next = Pj_Last + Pj_B - Pj_D
  I_Next = I_Last + I_B
  
  Pi_Last <- Pi_Next
  Pj_Last <- Pj_Next
  I_Last <- I_Next
  
  ## To prevent the populations from going negative
  if (Pi_Last < 0){
    Pi_Last <- 0
  }
  if (Pj_Last < 0){
    Pj_Last <- 0
  }
  ## To get host death if parasite abundance exceeds D threshold set by Antia
  if (Pi_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  if (Pj_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  
  return(c(Pi_Last, Pj_Last, I_Last))
}

DemStoch <- data.frame(time = seq(0,50,0.1))
SimNum <- seq(1,1000,1)
TT <- 50/0.1
for (i in 1:length(SimNum)){
  df <- data.frame()
  Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
  for (j in 1:TT){
    Output = DiscreteModel_Dem(Pi_Last, Pj_Last, I_Last)
    Pi_Last = Output[1]
    Pj_Last = Output[2]
    I_Last = Output[3]
    Addition <- c(Pi_Last, Pj_Last, I_Last)
    df <- rbind(df, Addition)
    colnames(df) <- c(paste0("Pi_", SimNum[i]),
                      paste0("Pj_", SimNum[i]),
                      paste0("I_", SimNum[i]))
  }
  inits <- data.frame(1, 0, 1)
  colnames(inits) <- c(paste0("Pi_", SimNum[i]),
                       paste0("Pj_", SimNum[i]),
                       paste0("I_", SimNum[i]))
  df <- data.frame(rbind(inits, df))
  DemStoch <- data.frame(cbind(DemStoch, df))
}
saveRDS(DemStoch, file = "Generated_Outputs/D_HT_Semi_1P.rds")

#### ... Pi-Advantage TPCs ####
DiscreteModel_Dem <- function(Pi_Last, Pj_Last, I_Last){
  ts = 0.1
  ri = 1.86; rj = 0; k = 0.01; p = 1; o = 1000
  
  Pi_B = rpois(1, ri*Pi_Last*ts) 
  Pi_D = rbinom(1, Pi_Last, (1-exp(-k*I_Last*ts))) 
  Pj_B = rpois(1, rj*Pj_Last*ts) 
  Pj_D = rbinom(1, Pj_Last, (1-exp(-k*I_Last*ts)))
  I_B = rpois(1, p*((Pi_Last/(Pi_Last + o)) + (Pj_Last/(Pj_Last + o)))*I_Last*ts) 
  
  Pi_Next = Pi_Last + Pi_B - Pi_D
  Pj_Next = Pj_Last + Pj_B - Pj_D
  I_Next = I_Last + I_B
  
  Pi_Last <- Pi_Next
  Pj_Last <- Pj_Next
  I_Last <- I_Next
  
  ## To prevent the populations from going negative
  if (Pi_Last < 0){
    Pi_Last <- 0
  }
  if (Pj_Last < 0){
    Pj_Last <- 0
  }
  ## To get host death if parasite abundance exceeds D threshold set by Antia
  if (Pi_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  if (Pj_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  
  return(c(Pi_Last, Pj_Last, I_Last))
}

DemStoch <- data.frame(time = seq(0,50,0.1))
SimNum <- seq(1,1000,1)
TT <- 50/0.1
for (i in 1:length(SimNum)){
  df <- data.frame()
  Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
  for (j in 1:TT){
    Output = DiscreteModel_Dem(Pi_Last, Pj_Last, I_Last)
    Pi_Last = Output[1]
    Pj_Last = Output[2]
    I_Last = Output[3]
    Addition <- c(Pi_Last, Pj_Last, I_Last)
    df <- rbind(df, Addition)
    colnames(df) <- c(paste0("Pi_", SimNum[i]),
                      paste0("Pj_", SimNum[i]),
                      paste0("I_", SimNum[i]))
  }
  inits <- data.frame(1, 0, 1)
  colnames(inits) <- c(paste0("Pi_", SimNum[i]),
                       paste0("Pj_", SimNum[i]),
                       paste0("I_", SimNum[i]))
  df <- data.frame(rbind(inits, df))
  DemStoch <- data.frame(cbind(DemStoch, df))
}
saveRDS(DemStoch, file = "Generated_Outputs/D_HT_PiAd_1P.rds")

#### ... Pj-Advantage, Version A TPCs ####
DiscreteModel_Dem <- function(Pi_Last, Pj_Last, I_Last){
  ts = 0.1
  ri = 1.705; rj = 0; k = 0.01; p = 1; o = 1000
  
  Pi_B = rpois(1, ri*Pi_Last*ts) 
  Pi_D = rbinom(1, Pi_Last, (1-exp(-k*I_Last*ts))) 
  Pj_B = rpois(1, rj*Pj_Last*ts) 
  Pj_D = rbinom(1, Pj_Last, (1-exp(-k*I_Last*ts)))
  I_B = rpois(1, p*((Pi_Last/(Pi_Last + o)) + (Pj_Last/(Pj_Last + o)))*I_Last*ts) 
  
  Pi_Next = Pi_Last + Pi_B - Pi_D
  Pj_Next = Pj_Last + Pj_B - Pj_D
  I_Next = I_Last + I_B
  
  Pi_Last <- Pi_Next
  Pj_Last <- Pj_Next
  I_Last <- I_Next
  
  ## To prevent the populations from going negative
  if (Pi_Last < 0){
    Pi_Last <- 0
  }
  if (Pj_Last < 0){
    Pj_Last <- 0
  }
  ## To get host death if parasite abundance exceeds D threshold set by Antia
  if (Pi_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  if (Pj_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  
  return(c(Pi_Last, Pj_Last, I_Last))
}

DemStoch <- data.frame(time = seq(0,50,0.1))
SimNum <- seq(1,1000,1)
TT <- 50/0.1
for (i in 1:length(SimNum)){
  df <- data.frame()
  Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
  for (j in 1:TT){
    Output = DiscreteModel_Dem(Pi_Last, Pj_Last, I_Last)
    Pi_Last = Output[1]
    Pj_Last = Output[2]
    I_Last = Output[3]
    Addition <- c(Pi_Last, Pj_Last, I_Last)
    df <- rbind(df, Addition)
    colnames(df) <- c(paste0("Pi_", SimNum[i]),
                      paste0("Pj_", SimNum[i]),
                      paste0("I_", SimNum[i]))
  }
  inits <- data.frame(1, 0, 1)
  colnames(inits) <- c(paste0("Pi_", SimNum[i]),
                       paste0("Pj_", SimNum[i]),
                       paste0("I_", SimNum[i]))
  df <- data.frame(rbind(inits, df))
  DemStoch <- data.frame(cbind(DemStoch, df))
}
saveRDS(DemStoch, file = "Generated_Outputs/D_HT_PjAd_1P.rds")

#### AUCs ####
TT <- seq(0, 50, 0.1)
Pi_Cols <- seq(2, 3001, 3); Pj_Cols <- seq(3, 3001, 3)

#### ... 1 parasite ####
# NOTE! If you want to use the outputs you've generated, replace "Outputs/TotalTransmission_Outputs" with "Generated_Outputs".
D_LT_Iden_1P <- readRDS("Outputs/TotalTransmission_Outputs/D_LT_Iden_1P.rds")
D_LT_Semi_1P <- readRDS("Outputs/TotalTransmission_Outputs/D_LT_Semi_1P.rds")
D_LT_PiAd_1P <- readRDS("Outputs/TotalTransmission_Outputs/D_LT_PiAd_1P.rds")
D_LT_PjAd_1P <- readRDS("Outputs/TotalTransmission_Outputs/D_LT_PjAd_1P.rds")
D_HT_Iden_1P <- readRDS("Outputs/TotalTransmission_Outputs/D_HT_Iden_1P.rds")
D_HT_Semi_1P <- readRDS("Outputs/TotalTransmission_Outputs/D_HT_Semi_1P.rds")
D_HT_PiAd_1P <- readRDS("Outputs/TotalTransmission_Outputs/D_HT_PiAd_1P.rds")
D_HT_PjAd_1P <- readRDS("Outputs/TotalTransmission_Outputs/D_HT_PjAd_1P.rds")

D_LT_Iden_1P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_LT_Iden_1P[ ,Pi_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
}
D_LT_Iden_1P_AUC <- cbind(D_LT_Iden_1P_AUC, Vec_Pi); D_LT_Iden_1P_AUC_Median <- median(D_LT_Iden_1P_AUC$Vec_Pi)

saveRDS(D_LT_Iden_1P_AUC, "Generated_Outputs/D_LT_Iden_1P_AUC.rds")

D_LT_Semi_1P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_LT_Semi_1P[ ,Pi_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
}
D_LT_Semi_1P_AUC <- cbind(D_LT_Semi_1P_AUC, Vec_Pi); D_LT_Semi_1P_AUC_Median <- median(D_LT_Semi_1P_AUC$Vec_Pi)

saveRDS(D_LT_Semi_1P_AUC, "Generated_Outputs/D_LT_Semi_1P_AUC.rds")

D_LT_PiAd_1P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_LT_PiAd_1P[ ,Pi_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
}
D_LT_PiAd_1P_AUC <- cbind(D_LT_PiAd_1P_AUC, Vec_Pi); D_LT_PiAd_1P_AUC_Median <- median(D_LT_PiAd_1P_AUC$Vec_Pi)

saveRDS(D_LT_PiAd_1P_AUC, "Generated_Outputs/D_LT_PiAd_1P_AUC.rds")

D_LT_PjAd_1P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_LT_PjAd_1P[ ,Pi_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
}
D_LT_PjAd_1P_AUC <- cbind(D_LT_PjAd_1P_AUC, Vec_Pi); D_LT_PjAd_1P_AUC_Median <- median(D_LT_PjAd_1P_AUC$Vec_Pi)

saveRDS(D_LT_PjAd_1P_AUC, "Generated_Outputs/D_LT_PjAd_1P_AUC.rds")

D_HT_Iden_1P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_HT_Iden_1P[ ,Pi_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
}
D_HT_Iden_1P_AUC <- cbind(D_HT_Iden_1P_AUC, Vec_Pi); D_HT_Iden_1P_AUC_Median <- median(D_HT_Iden_1P_AUC$Vec_Pi)

saveRDS(D_HT_Iden_1P_AUC, "Generated_Outputs/D_HT_Iden_1P_AUC.rds")

D_HT_Semi_1P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_HT_Semi_1P[ ,Pi_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
}
D_HT_Semi_1P_AUC <- cbind(D_HT_Semi_1P_AUC, Vec_Pi); D_HT_Semi_1P_AUC_Median <- median(D_HT_Semi_1P_AUC$Vec_Pi)

saveRDS(D_HT_Semi_1P_AUC, "Generated_Outputs/D_HT_Semi_1P_AUC.rds")

D_HT_PiAd_1P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_HT_PiAd_1P[ ,Pi_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
}
D_HT_PiAd_1P_AUC <- cbind(D_HT_PiAd_1P_AUC, Vec_Pi); D_HT_PiAd_1P_AUC_Median <- median(D_HT_PiAd_1P_AUC$Vec_Pi)

saveRDS(D_HT_PiAd_1P_AUC, "Generated_Outputs/D_HT_PiAd_1P_AUC.rds")

D_HT_PjAd_1P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_HT_PjAd_1P[ ,Pi_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
}
D_HT_PjAd_1P_AUC <- cbind(D_HT_PjAd_1P_AUC, Vec_Pi); D_HT_PjAd_1P_AUC_Median <- median(D_HT_PjAd_1P_AUC$Vec_Pi)

saveRDS(D_HT_PjAd_1P_AUC, "Generated_Outputs/D_HT_PjAd_1P_AUC.rds")

#### ... 2 parasites ####
# NOTE! If you want to use the outputs you've generated, replace "Data/DemStoch_Outputs" with "Generated_Outputs".
D_LT_Iden <- readRDS("Data/DemStoch_Outputs/D_LT_Iden.rds"); D_LT_Iden_2P <- D_LT_Iden[[1]]
D_LT_Semi <- readRDS("Data/DemStoch_Outputs/D_LT_Semi.rds"); D_LT_Semi_2P <- D_LT_Semi[[1]]
D_LT_PiAd <- readRDS("Data/DemStoch_Outputs/D_LT_PiAd.rds"); D_LT_PiAd_2P <- D_LT_PiAd[[1]]
D_LT_PjAd <- readRDS("Data/DemStoch_Outputs/D_LT_PjAd.rds"); D_LT_PjAd_2P <- D_LT_PjAd[[1]]
D_HT_Iden <- readRDS("Data/DemStoch_Outputs/D_HT_Iden.rds"); D_HT_Iden_2P <- D_LT_Iden[[1]]
D_HT_Semi <- readRDS("Data/DemStoch_Outputs/D_HT_Semi.rds"); D_HT_Semi_2P <- D_LT_Semi[[1]]
D_HT_PiAd <- readRDS("Data/DemStoch_Outputs/D_HT_PiAd.rds"); D_HT_PiAd_2P <- D_LT_PiAd[[1]]
D_HT_PjAd <- readRDS("Data/DemStoch_Outputs/D_HT_PjAd.rds"); D_HT_PjAd_2P <- D_LT_PjAd[[1]]

D_LT_Iden_2P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c(); Vec_Pj <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_LT_Iden_2P[ ,Pi_Cols[j]]
  Pj <- D_LT_Iden_2P[ ,Pj_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
  Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
}
D_LT_Iden_2P_AUC <- cbind(D_LT_Iden_2P_AUC, Vec_Pi, Vec_Pj)
D_LT_Iden_2P_AUC_PiMedian <- median(D_LT_Iden_2P_AUC$Vec_Pi); D_LT_Iden_2P_AUC_PjMedian <- median(D_LT_Iden_2P_AUC$Vec_Pj)

saveRDS(D_LT_Iden_2P_AUC, "Generated_Outputs/D_LT_Iden_2P_AUC.rds")

D_LT_Semi_2P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c(); Vec_Pj <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_LT_Semi_2P[ ,Pi_Cols[j]]
  Pj <- D_LT_Semi_2P[ ,Pj_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
  Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
}
D_LT_Semi_2P_AUC <- cbind(D_LT_Semi_2P_AUC, Vec_Pi, Vec_Pj)
D_LT_Semi_2P_AUC_PiMedian <- median(D_LT_Semi_2P_AUC$Vec_Pi); D_LT_Semi_2P_AUC_PjMedian <- median(D_LT_Semi_2P_AUC$Vec_Pj)

saveRDS(D_LT_Semi_2P_AUC, "Generated_Outputs/D_LT_Semi_2P_AUC.rds")

D_LT_PiAd_2P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c(); Vec_Pj <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_LT_PiAd_2P[ ,Pi_Cols[j]]
  Pj <- D_LT_PiAd_2P[ ,Pj_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
  Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
}
D_LT_PiAd_2P_AUC <- cbind(D_LT_PiAd_2P_AUC, Vec_Pi, Vec_Pj)
D_LT_PiAd_2P_AUC_PiMedian <- median(D_LT_PiAd_2P_AUC$Vec_Pi); D_LT_PiAd_2P_AUC_PjMedian <- median(D_LT_PiAd_2P_AUC$Vec_Pj)

saveRDS(D_LT_PiAd_2P_AUC, "Generated_Outputs/D_LT_PiAd_2P_AUC.rds")

D_LT_PjAd_2P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c(); Vec_Pj <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_LT_PjAd_2P[ ,Pi_Cols[j]]
  Pj <- D_LT_PjAd_2P[ ,Pj_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
  Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
}
D_LT_PjAd_2P_AUC <- cbind(D_LT_PjAd_2P_AUC, Vec_Pi, Vec_Pj)
D_LT_PjAd_2P_AUC_PiMedian <- median(D_LT_PjAd_2P_AUC$Vec_Pi); D_LT_PjAd_2P_AUC_PjMedian <- median(D_LT_PjAd_2P_AUC$Vec_Pj)

saveRDS(D_LT_PjAd_2P_AUC, "Generated_Outputs/D_LT_PjAd_2P_AUC.rds")

D_HT_Iden_2P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c(); Vec_Pj <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_HT_Iden_2P[ ,Pi_Cols[j]]
  Pj <- D_HT_Iden_2P[ ,Pj_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
  Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
}
D_HT_Iden_2P_AUC <- cbind(D_HT_Iden_2P_AUC, Vec_Pi, Vec_Pj)
D_HT_Iden_2P_AUC_PiMedian <- median(D_HT_Iden_2P_AUC$Vec_Pi); D_HT_Iden_2P_AUC_PjMedian <- median(D_HT_Iden_2P_AUC$Vec_Pj)

saveRDS(D_HT_Iden_2P_AUC, "Generated_Outputs/D_HT_Iden_2P_AUC.rds")

D_HT_Semi_2P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c(); Vec_Pj <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_HT_Semi_2P[ ,Pi_Cols[j]]
  Pj <- D_HT_Semi_2P[ ,Pj_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
  Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
}
D_HT_Semi_2P_AUC <- cbind(D_HT_Semi_2P_AUC, Vec_Pi, Vec_Pj)
D_HT_Semi_2P_AUC_PiMedian <- median(D_HT_Semi_2P_AUC$Vec_Pi); D_HT_Semi_2P_AUC_PjMedian <- median(D_HT_Semi_2P_AUC$Vec_Pj)

saveRDS(D_HT_Semi_2P_AUC, "Generated_Outputs/D_HT_Semi_2P_AUC.rds")

D_HT_PiAd_2P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c(); Vec_Pj <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_HT_PiAd_2P[ ,Pi_Cols[j]]
  Pj <- D_HT_PiAd_2P[ ,Pj_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
  Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
}
D_HT_PiAd_2P_AUC <- cbind(D_HT_PiAd_2P_AUC, Vec_Pi, Vec_Pj)
D_HT_PiAd_2P_AUC_PiMedian <- median(D_HT_PiAd_2P_AUC$Vec_Pi); D_HT_PiAd_2P_AUC_PjMedian <- median(D_HT_PiAd_2P_AUC$Vec_Pj)

saveRDS(D_HT_PiAd_2P_AUC, "Generated_Outputs/D_HT_PiAd_2P_AUC.rds")

D_HT_PjAd_2P_AUC <- data.frame(Sim = seq(1, 1000, 1))
Vec_Pi <- c(); Vec_Pj <- c()
for (j in 1:length(Pi_Cols)){
  Pi <- D_HT_PjAd_2P[ ,Pi_Cols[j]]
  Pj <- D_HT_PjAd_2P[ ,Pj_Cols[j]]
  Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline")
  Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
}
D_HT_PjAd_2P_AUC <- cbind(D_HT_PjAd_2P_AUC, Vec_Pi, Vec_Pj)
D_HT_PjAd_2P_AUC_PiMedian <- median(D_HT_PjAd_2P_AUC$Vec_Pi); D_HT_PjAd_2P_AUC_PjMedian <- median(D_HT_PjAd_2P_AUC$Vec_Pj)

saveRDS(D_HT_PjAd_2P_AUC, "Generated_Outputs/D_HT_PjAd_2P_AUC.rds")
