==> 6_Transmisson_SingleParasite_Sim.R <==
#### DOC STRING ===============================================================
# This script runs demographically and environmentally stochastic single-parasite 
# model simulations.

# If you choose to run this script, outputs will be stored in 
# "EnvironmentalVariability_Repo/Generated_Outputs".

# To see the outputs generated by this script without running it, go to 
# "EnvironmentalVariability_Repo/Outputs/TotalTransmission_Outputs".
#### END DOC STRING ===========================================================

# Required packages
install.packages("dplyr"); library(dplyr)

#### DemEnvStoch Model ####
DemEnvStoch_Model <- function(Pi_Last, Pj_Last, I_Last, ri, rj){
  ts = 0.1
  k = 0.01; p = 1; o = 1000
  
  Pi_B = rpois(1, ri*Pi_Last*ts) 
  Pi_D = rbinom(1, Pi_Last, (1-exp(-k*I_Last*ts))) 
  Pj_B = rpois(1, rj*Pj_Last*ts) 
  Pj_D = rbinom(1, Pj_Last, (1-exp(-k*I_Last*ts)))
  I_B = rpois(1, p*((Pi_Last/(Pi_Last + o)) + (Pj_Last/(Pj_Last + o)))*I_Last*ts) 
  
  Pi_Next = Pi_Last + Pi_B - Pi_D
  Pj_Next = Pj_Last + Pj_B - Pj_D
  I_Next = I_Last + I_B
  
  Pi_Last <- Pi_Next
  Pj_Last <- Pj_Next
  I_Last <- I_Next
  
  ## To get pop'ns to 0
  if (Pi_Last < 1){
    Pi_Last <- 0
  }
  if (Pj_Last < 1){
    Pj_Last <- 0
  }
  
  ## To get host death if parasite abundance exceeds D threshold set by Antia
  if (Pi_Last > 10^9 || Pj_Last > 10^9){
    Pi_Last <- 0
    Pj_Last <- 0
    I_Last <- 0
  }
  
  return(c(Pi_Last, Pj_Last, I_Last))
}

#### Low-Average Temperature Environment ####
L <- 499
ar <- 0.25
K <- 273.15
range <- 30

freq <- c(1, 2, 5, seq(10, 100, 10))
IntroPt <- c(0)
sd <- seq(2, range, 2)
SimNum <- seq(1, 1000, 1)
TT <- round(seq(0, 49.9, 0.1), 2)

#### ... Identical TPCs ####
MTE_TPC_Pi <- function(temp){
  lambda_knot <- 1.4; Tknot <- 14+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-9.5)+K; TzH <- 22.5+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}
MTE_TPC_Pj <- function(temp){
  lambda_knot <- 1.4; Tknot <- 14+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-9.5)+K; TzH <- 22.5+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}

# Temp Sampling Objects
av <- 19.9

#### ... ... Simulation ####
List3 <- vector(mode = "list", length = length(freq))
for (i in 1:length(freq)){ 
  List2 <- vector(mode = "list", length = length(IntroPt))
  for (j in 1:length(IntroPt)){ 
    List1 <- vector(mode = "list", length = length(sd))
    for (k in 1:length(sd)){ 
      SD_df <- data.frame(time = seq(0,50,0.1))
      for (l in 1:length(SimNum)){
        temp_sequence <- c()
        temp_sequence[1] <- 0
        for (m in 1:L) {
          temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)
          while(temp_sequence[m + 1] + av <= 0 | temp_sequence[m + 1] + av >= range)
          {temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)}
        }
        temp_sequence <- temp_sequence + av
        temp_sequence <- temp_sequence + K # Kelvin
        temp_sequence <- rep(temp_sequence, each = freq[i]) 
        temp_sequence <- temp_sequence[-(length(TT)+1:length(temp_sequence))]
        ri <- MTE_TPC_Pi(temp_sequence)
        rj <- MTE_TPC_Pj(temp_sequence)
        
        df <- data.frame()
        Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
        for (n in 1:length(TT)){
          
          if (TT[n] == IntroPt[j]){
            Pj_Last <- 0
          }
          
          Output = DemEnvStoch_Model(Pi_Last, Pj_Last, I_Last, ri[n], rj[n])
          Pi_Last = Output[1]
          Pj_Last = Output[2]
          I_Last = Output[3]
          Addition <- c(Pi_Last, Pj_Last, I_Last)
          df <- rbind(df, Addition)
          colnames(df) <- c(paste0("Pi_", SimNum[l]),
                            paste0("Pj_", SimNum[l]),
                            paste0("I_", SimNum[l]))
        }
        Inits <- data.frame(1, 0, 1)
        if (IntroPt[j] == 0){
          Inits <- data.frame(1, 0, 1)
        }
        colnames(Inits) <- c(paste0("Pi_", SimNum[l]),
                             paste0("Pj_", SimNum[l]),
                             paste0("I_", SimNum[l]))
        df <- data.frame(rbind(Inits, df))
        SD_df <- data.frame(cbind(SD_df, df))
      }
      List1[[k]] <- SD_df
    }
    List1Names <- c()
    for (o in 1:length(sd)){
      List1Names[o] <- paste0("SD_", sd[o])
    }
    names(List1) <- List1Names
    List2[[j]] <- List1
  }
  List2Names <- c()
  for (p in 1:length(IntroPt)){
    List2Names[p] <- paste0("Intro_", IntroPt[p])
  }
  names(List2) <- List2Names
  List3[[i]] <- List2
}
List3Names <- c()
for (i in 1:length(freq)){
  List3Names[i] <- paste0("Each_", freq[i])
}
names(List3) <- List3Names

saveRDS(List3, file = "Generated_Outputs/DE_LT_Iden_1P.rds")

#### ... Semi-Identical TPCs ####
MTE_TPC_Pi <- function(temp){
  lambda_knot <- 1.5; Tknot <- 13+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-12)+K; TzH <- 20.6+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}
MTE_TPC_Pj <- function(temp){
  lambda_knot <- 1.54; Tknot <- 18+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-7)+K; TzH <- 25.6+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}

# Temp Sampling Objects
av <- 19.9

#### ... ... Simulation ####
List3 <- vector(mode = "list", length = length(freq))
for (i in 1:length(freq)){ 
  List2 <- vector(mode = "list", length = length(IntroPt))
  for (j in 1:length(IntroPt)){ 
    List1 <- vector(mode = "list", length = length(sd))
    for (k in 1:length(sd)){ 
      SD_df <- data.frame(time = seq(0,50,0.1))
      for (l in 1:length(SimNum)){
        temp_sequence <- c()
        temp_sequence[1] <- 0
        for (m in 1:L) {
          temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)
          while(temp_sequence[m + 1] + av <= 0 | temp_sequence[m + 1] + av >= range)
          {temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)}
        }
        temp_sequence <- temp_sequence + av
        temp_sequence <- temp_sequence + K # Kelvin
        temp_sequence <- rep(temp_sequence, each = freq[i]) 
        temp_sequence <- temp_sequence[-(length(TT)+1:length(temp_sequence))]
        ri <- MTE_TPC_Pi(temp_sequence)
        rj <- MTE_TPC_Pj(temp_sequence)
        
        df <- data.frame()
        Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
        for (n in 1:length(TT)){
          
          if (TT[n] == IntroPt[j]){
            Pj_Last <- 0
          }
          
          Output = DemEnvStoch_Model(Pi_Last, Pj_Last, I_Last, ri[n], rj[n])
          Pi_Last = Output[1]
          Pj_Last = Output[2]
          I_Last = Output[3]
          Addition <- c(Pi_Last, Pj_Last, I_Last)
          df <- rbind(df, Addition)
          colnames(df) <- c(paste0("Pi_", SimNum[l]),
                            paste0("Pj_", SimNum[l]),
                            paste0("I_", SimNum[l]))
        }
        Inits <- data.frame(1, 0, 1)
        if (IntroPt[j] == 0){
          Inits <- data.frame(1, 0, 1)
        }
        colnames(Inits) <- c(paste0("Pi_", SimNum[l]),
                             paste0("Pj_", SimNum[l]),
                             paste0("I_", SimNum[l]))
        df <- data.frame(rbind(Inits, df))
        SD_df <- data.frame(cbind(SD_df, df))
      }
      List1[[k]] <- SD_df
    }
    List1Names <- c()
    for (o in 1:length(sd)){
      List1Names[o] <- paste0("SD_", sd[o])
    }
    names(List1) <- List1Names
    List2[[j]] <- List1
  }
  List2Names <- c()
  for (p in 1:length(IntroPt)){
    List2Names[p] <- paste0("Intro_", IntroPt[p])
  }
  names(List2) <- List2Names
  List3[[i]] <- List2
}
List3Names <- c()
for (i in 1:length(freq)){
  List3Names[i] <- paste0("Each_", freq[i])
}
names(List3) <- List3Names

saveRDS(List3, file = "Generated_Outputs/DE_LT_Semi_1P.rds")

#### ... Pi-Advantage TPCs ####
MTE_TPC_Pi <- function(temp){
  lambda_knot <- 1.5; Tknot <- 13+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-12)+K; TzH <- 20.6+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}
MTE_TPC_Pj <- function(temp){
  lambda_knot <- 1.54; Tknot <- 18+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-7)+K; TzH <- 25.6+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}

# Temp Sampling Objects
av <- 18

#### ... ... Simulation ####
List3 <- vector(mode = "list", length = length(freq))
for (i in 1:length(freq)){ 
  List2 <- vector(mode = "list", length = length(IntroPt))
  for (j in 1:length(IntroPt)){ 
    List1 <- vector(mode = "list", length = length(sd))
    for (k in 1:length(sd)){ 
      SD_df <- data.frame(time = seq(0,50,0.1))
      for (l in 1:length(SimNum)){
        temp_sequence <- c()
        temp_sequence[1] <- 0
        for (m in 1:L) {
          temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)
          while(temp_sequence[m + 1] + av <= 0 | temp_sequence[m + 1] + av >= range)
          {temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)}
        }
        temp_sequence <- temp_sequence + av
        temp_sequence <- temp_sequence + K # Kelvin
        temp_sequence <- rep(temp_sequence, each = freq[i]) 
        temp_sequence <- temp_sequence[-(length(TT)+1:length(temp_sequence))]
        ri <- MTE_TPC_Pi(temp_sequence)
        rj <- MTE_TPC_Pj(temp_sequence)
        
        df <- data.frame()
        Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
        for (n in 1:length(TT)){
          
          if (TT[n] == IntroPt[j]){
            Pj_Last <- 0
          }
          
          Output = DemEnvStoch_Model(Pi_Last, Pj_Last, I_Last, ri[n], rj[n])
          Pi_Last = Output[1]
          Pj_Last = Output[2]
          I_Last = Output[3]
          Addition <- c(Pi_Last, Pj_Last, I_Last)
          df <- rbind(df, Addition)
          colnames(df) <- c(paste0("Pi_", SimNum[l]),
                            paste0("Pj_", SimNum[l]),
                            paste0("I_", SimNum[l]))
        }
        Inits <- data.frame(1, 0, 1)
        if (IntroPt[j] == 0){
          Inits <- data.frame(1, 0, 1)
        }
        colnames(Inits) <- c(paste0("Pi_", SimNum[l]),
                             paste0("Pj_", SimNum[l]),
                             paste0("I_", SimNum[l]))
        df <- data.frame(rbind(Inits, df))
        SD_df <- data.frame(cbind(SD_df, df))
      }
      List1[[k]] <- SD_df
    }
    List1Names <- c()
    for (o in 1:length(sd)){
      List1Names[o] <- paste0("SD_", sd[o])
    }
    names(List1) <- List1Names
    List2[[j]] <- List1
  }
  List2Names <- c()
  for (p in 1:length(IntroPt)){
    List2Names[p] <- paste0("Intro_", IntroPt[p])
  }
  names(List2) <- List2Names
  List3[[i]] <- List2
}
List3Names <- c()
for (i in 1:length(freq)){
  List3Names[i] <- paste0("Each_", freq[i])
}
names(List3) <- List3Names

saveRDS(List3, file = "Generated_Outputs/DE_LT_PiAd_1P.rds")

#### ... Pj-Advantage, Version A TPCs ####
MTE_TPC_Pi <- function(temp){
  lambda_knot <- 1.5; Tknot <- 13+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-12)+K; TzH <- 20.6+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}
MTE_TPC_Pj <- function(temp){
  lambda_knot <- 1.54; Tknot <- 18+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-7)+K; TzH <- 25.6+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}

# Temp Sampling Objects
av <- 20.8530117

#### ... ... Simulation ####
List3 <- vector(mode = "list", length = length(freq))
for (i in 1:length(freq)){ 
  List2 <- vector(mode = "list", length = length(IntroPt))
  for (j in 1:length(IntroPt)){ 
    List1 <- vector(mode = "list", length = length(sd))
    for (k in 1:length(sd)){ 
      SD_df <- data.frame(time = seq(0,50,0.1))
      for (l in 1:length(SimNum)){
        temp_sequence <- c()
        temp_sequence[1] <- 0
        for (m in 1:L) {
          temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)
          while(temp_sequence[m + 1] + av <= 0 | temp_sequence[m + 1] + av >= range)
          {temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)}
        }
        temp_sequence <- temp_sequence + av
        temp_sequence <- temp_sequence + K # Kelvin
        temp_sequence <- rep(temp_sequence, each = freq[i]) 
        temp_sequence <- temp_sequence[-(length(TT)+1:length(temp_sequence))]
        ri <- MTE_TPC_Pi(temp_sequence)
        rj <- MTE_TPC_Pj(temp_sequence)
        
        df <- data.frame()
        Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
        for (n in 1:length(TT)){
          
          if (TT[n] == IntroPt[j]){
            Pj_Last <- 0
          }
          
          Output = DemEnvStoch_Model(Pi_Last, Pj_Last, I_Last, ri[n], rj[n])
          Pi_Last = Output[1]
          Pj_Last = Output[2]
          I_Last = Output[3]
          Addition <- c(Pi_Last, Pj_Last, I_Last)
          df <- rbind(df, Addition)
          colnames(df) <- c(paste0("Pi_", SimNum[l]),
                            paste0("Pj_", SimNum[l]),
                            paste0("I_", SimNum[l]))
        }
        Inits <- data.frame(1, 0, 1)
        if (IntroPt[j] == 0){
          Inits <- data.frame(1, 0, 1)
        }
        colnames(Inits) <- c(paste0("Pi_", SimNum[l]),
                             paste0("Pj_", SimNum[l]),
                             paste0("I_", SimNum[l]))
        df <- data.frame(rbind(Inits, df))
        SD_df <- data.frame(cbind(SD_df, df))
      }
      List1[[k]] <- SD_df
    }
    List1Names <- c()
    for (o in 1:length(sd)){
      List1Names[o] <- paste0("SD_", sd[o])
    }
    names(List1) <- List1Names
    List2[[j]] <- List1
  }
  List2Names <- c()
  for (p in 1:length(IntroPt)){
    List2Names[p] <- paste0("Intro_", IntroPt[p])
  }
  names(List2) <- List2Names
  List3[[i]] <- List2
}
List3Names <- c()
for (i in 1:length(freq)){
  List3Names[i] <- paste0("Each_", freq[i])
}
names(List3) <- List3Names

saveRDS(List3, file = "Generated_Outputs/DE_LT_PjAd_1P.rds")

#### High-Average Temperature Environment ####
L <- 499
ar <- 0.25
K <- 273.15
range <- 40

freq <- c(1, 2, 5, seq(10, 100, 10))
IntroPt <- c(0)
sd <- seq(2, range, 2)
SimNum <- seq(1, 1000, 1)
TT <- round(seq(0, 49.9, 0.1), 2)

#### ... Identical TPCs ####
MTE_TPC_Pi <- function(temp){
  lambda_knot <- 1.2; Tknot <- 19+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-2)+K; TzH <- 29.8+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}
MTE_TPC_Pj <- function(temp){
  lambda_knot <- 1.2; Tknot <- 19+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-2)+K; TzH <- 29.8+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}

# Temp Sampling Objects
av <- 27.1

#### ... ... Simulation ####
List3 <- vector(mode = "list", length = length(freq))
for (i in 1:length(freq)){ 
  List2 <- vector(mode = "list", length = length(IntroPt))
  for (j in 1:length(IntroPt)){ 
    List1 <- vector(mode = "list", length = length(sd))
    for (k in 1:length(sd)){ 
      SD_df <- data.frame(time = seq(0,50,0.1))
      for (l in 1:length(SimNum)){
        temp_sequence <- c()
        temp_sequence[1] <- 0
        for (m in 1:L) {
          temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)
          while(temp_sequence[m + 1] + av <= 0 | temp_sequence[m + 1] + av >= range)
          {temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)}
        }
        temp_sequence <- temp_sequence + av
        temp_sequence <- temp_sequence + K # Kelvin
        temp_sequence <- rep(temp_sequence, each = freq[i]) 
        temp_sequence <- temp_sequence[-(length(TT)+1:length(temp_sequence))]
        ri <- MTE_TPC_Pi(temp_sequence)
        rj <- MTE_TPC_Pj(temp_sequence)
        
        df <- data.frame()
        Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
        for (n in 1:length(TT)){
          
          if (TT[n] == IntroPt[j]){
            Pj_Last <- 0
          }
          
          Output = DemEnvStoch_Model(Pi_Last, Pj_Last, I_Last, ri[n], rj[n])
          Pi_Last = Output[1]
          Pj_Last = Output[2]
          I_Last = Output[3]
          Addition <- c(Pi_Last, Pj_Last, I_Last)
          df <- rbind(df, Addition)
          colnames(df) <- c(paste0("Pi_", SimNum[l]),
                            paste0("Pj_", SimNum[l]),
                            paste0("I_", SimNum[l]))
        }
        Inits <- data.frame(1, 0, 1)
        if (IntroPt[j] == 0){
          Inits <- data.frame(1, 0, 1)
        }
        colnames(Inits) <- c(paste0("Pi_", SimNum[l]),
                             paste0("Pj_", SimNum[l]),
                             paste0("I_", SimNum[l]))
        df <- data.frame(rbind(Inits, df))
        SD_df <- data.frame(cbind(SD_df, df))
      }
      List1[[k]] <- SD_df
    }
    List1Names <- c()
    for (o in 1:length(sd)){
      List1Names[o] <- paste0("SD_", sd[o])
    }
    names(List1) <- List1Names
    List2[[j]] <- List1
  }
  List2Names <- c()
  for (p in 1:length(IntroPt)){
    List2Names[p] <- paste0("Intro_", IntroPt[p])
  }
  names(List2) <- List2Names
  List3[[i]] <- List2
}
List3Names <- c()
for (i in 1:length(freq)){
  List3Names[i] <- paste0("Each_", freq[i])
}
names(List3) <- List3Names

saveRDS(List3, file = "Generated_Outputs/DE_HT_Iden_1P.rds")

#### ... Semi-Identical TPCs ####
MTE_TPC_Pi <- function(temp){
  lambda_knot <- 0.68; Tknot <- 11.3+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-10)+K; TzH <- 28+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}
MTE_TPC_Pj <- function(temp){
  lambda_knot <- 1.68; Tknot <- 26+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- 1+K; TzH <- 33+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}

# Temp Sampling Objects
av <- 27.1

#### ... ... Simulation ####
List3 <- vector(mode = "list", length = length(freq))
for (i in 1:length(freq)){ 
  List2 <- vector(mode = "list", length = length(IntroPt))
  for (j in 1:length(IntroPt)){ 
    List1 <- vector(mode = "list", length = length(sd))
    for (k in 1:length(sd)){ 
      SD_df <- data.frame(time = seq(0,50,0.1))
      for (l in 1:length(SimNum)){
        temp_sequence <- c()
        temp_sequence[1] <- 0
        for (m in 1:L) {
          temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)
          while(temp_sequence[m + 1] + av <= 0 | temp_sequence[m + 1] + av >= range)
          {temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)}
        }
        temp_sequence <- temp_sequence + av
        temp_sequence <- temp_sequence + K # Kelvin
        temp_sequence <- rep(temp_sequence, each = freq[i]) 
        temp_sequence <- temp_sequence[-(length(TT)+1:length(temp_sequence))]
        ri <- MTE_TPC_Pi(temp_sequence)
        rj <- MTE_TPC_Pj(temp_sequence)
        
        df <- data.frame()
        Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
        for (n in 1:length(TT)){
          
          if (TT[n] == IntroPt[j]){
            Pj_Last <- 0
          }
          
          Output = DemEnvStoch_Model(Pi_Last, Pj_Last, I_Last, ri[n], rj[n])
          Pi_Last = Output[1]
          Pj_Last = Output[2]
          I_Last = Output[3]
          Addition <- c(Pi_Last, Pj_Last, I_Last)
          df <- rbind(df, Addition)
          colnames(df) <- c(paste0("Pi_", SimNum[l]),
                            paste0("Pj_", SimNum[l]),
                            paste0("I_", SimNum[l]))
        }
        Inits <- data.frame(1, 0, 1)
        if (IntroPt[j] == 0){
          Inits <- data.frame(1, 0, 1)
        }
        colnames(Inits) <- c(paste0("Pi_", SimNum[l]),
                             paste0("Pj_", SimNum[l]),
                             paste0("I_", SimNum[l]))
        df <- data.frame(rbind(Inits, df))
        SD_df <- data.frame(cbind(SD_df, df))
      }
      List1[[k]] <- SD_df
    }
    List1Names <- c()
    for (o in 1:length(sd)){
      List1Names[o] <- paste0("SD_", sd[o])
    }
    names(List1) <- List1Names
    List2[[j]] <- List1
  }
  List2Names <- c()
  for (p in 1:length(IntroPt)){
    List2Names[p] <- paste0("Intro_", IntroPt[p])
  }
  names(List2) <- List2Names
  List3[[i]] <- List2
}
List3Names <- c()
for (i in 1:length(freq)){
  List3Names[i] <- paste0("Each_", freq[i])
}
names(List3) <- List3Names

saveRDS(List3, file = "Generated_Outputs/DE_HT_Semi_1P.rds")

#### ... Pi-Advantage TPCs ####
MTE_TPC_Pi <- function(temp){
  lambda_knot <- 0.68; Tknot <- 11.3+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-10)+K; TzH <- 28+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}
MTE_TPC_Pj <- function(temp){
  lambda_knot <- 1.68; Tknot <- 26+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- 1+K; TzH <- 33+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}

# Temp Sampling Objects
av <- 25.5

#### ... ... Simulation ####
List3 <- vector(mode = "list", length = length(freq))
for (i in 1:length(freq)){ 
  List2 <- vector(mode = "list", length = length(IntroPt))
  for (j in 1:length(IntroPt)){ 
    List1 <- vector(mode = "list", length = length(sd))
    for (k in 1:length(sd)){ 
      SD_df <- data.frame(time = seq(0,50,0.1))
      for (l in 1:length(SimNum)){
        temp_sequence <- c()
        temp_sequence[1] <- 0
        for (m in 1:L) {
          temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)
          while(temp_sequence[m + 1] + av <= 0 | temp_sequence[m + 1] + av >= range)
          {temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)}
        }
        temp_sequence <- temp_sequence + av
        temp_sequence <- temp_sequence + K # Kelvin
        temp_sequence <- rep(temp_sequence, each = freq[i]) 
        temp_sequence <- temp_sequence[-(length(TT)+1:length(temp_sequence))]
        ri <- MTE_TPC_Pi(temp_sequence)
        rj <- MTE_TPC_Pj(temp_sequence)
        
        df <- data.frame()
        Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
        for (n in 1:length(TT)){
          
          if (TT[n] == IntroPt[j]){
            Pj_Last <- 0
          }
          
          Output = DemEnvStoch_Model(Pi_Last, Pj_Last, I_Last, ri[n], rj[n])
          Pi_Last = Output[1]
          Pj_Last = Output[2]
          I_Last = Output[3]
          Addition <- c(Pi_Last, Pj_Last, I_Last)
          df <- rbind(df, Addition)
          colnames(df) <- c(paste0("Pi_", SimNum[l]),
                            paste0("Pj_", SimNum[l]),
                            paste0("I_", SimNum[l]))
        }
        Inits <- data.frame(1, 0, 1)
        if (IntroPt[j] == 0){
          Inits <- data.frame(1, 0, 1)
        }
        colnames(Inits) <- c(paste0("Pi_", SimNum[l]),
                             paste0("Pj_", SimNum[l]),
                             paste0("I_", SimNum[l]))
        df <- data.frame(rbind(Inits, df))
        SD_df <- data.frame(cbind(SD_df, df))
      }
      List1[[k]] <- SD_df
    }
    List1Names <- c()
    for (o in 1:length(sd)){
      List1Names[o] <- paste0("SD_", sd[o])
    }
    names(List1) <- List1Names
    List2[[j]] <- List1
  }
  List2Names <- c()
  for (p in 1:length(IntroPt)){
    List2Names[p] <- paste0("Intro_", IntroPt[p])
  }
  names(List2) <- List2Names
  List3[[i]] <- List2
}
List3Names <- c()
for (i in 1:length(freq)){
  List3Names[i] <- paste0("Each_", freq[i])
}
names(List3) <- List3Names

saveRDS(List3, file = "Generated_Outputs/DE_HT_PiAd_1P.rds")

#### ... Pj-Advantage, Version A TPCs ####
MTE_TPC_Pi <- function(temp){
  lambda_knot <- 0.68; Tknot <- 11.3+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- (-10)+K; TzH <- 28+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}
MTE_TPC_Pj <- function(temp){
  lambda_knot <- 1.68; Tknot <- 26+K; Ez <- 0.65; EzL <- (-6.2); EzH <- 6.2; TzL <- 1+K; TzH <- 33+K; k <- 8.62*10^(-5)
  r <- lambda_knot*
    (exp(-(Ez/k)*((1/temp) - (1/Tknot))))*
    (1 + exp((EzL/k)*((1/TzL) - (1/temp))) + exp((EzH/k)*((1/TzH) - (1/temp))))^(-1)
  return(r)
}

# Temp Sampling Objects
av <- 28.0639755

#### ... ... Simulation ####
List3 <- vector(mode = "list", length = length(freq))
for (i in 1:length(freq)){ 
  List2 <- vector(mode = "list", length = length(IntroPt))
  for (j in 1:length(IntroPt)){ 
    List1 <- vector(mode = "list", length = length(sd))
    for (k in 1:length(sd)){ 
      SD_df <- data.frame(time = seq(0,50,0.1))
      for (l in 1:length(SimNum)){
        temp_sequence <- c()
        temp_sequence[1] <- 0
        for (m in 1:L) {
          temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)
          while(temp_sequence[m + 1] + av <= 0 | temp_sequence[m + 1] + av >= range)
          {temp_sequence[m + 1] <- round(ar * temp_sequence[m] + rnorm(1, 0, sd[k]), digits = 0)}
        }
        temp_sequence <- temp_sequence + av
        temp_sequence <- temp_sequence + K # Kelvin
        temp_sequence <- rep(temp_sequence, each = freq[i]) 
        temp_sequence <- temp_sequence[-(length(TT)+1:length(temp_sequence))]
        ri <- MTE_TPC_Pi(temp_sequence)
        rj <- MTE_TPC_Pj(temp_sequence)
        
        df <- data.frame()
        Pi_Last <- 1; Pj_Last <- 0; I_Last <- 1
        for (n in 1:length(TT)){
          
          if (TT[n] == IntroPt[j]){
            Pj_Last <- 0
          }
          
          Output = DemEnvStoch_Model(Pi_Last, Pj_Last, I_Last, ri[n], rj[n])
          Pi_Last = Output[1]
          Pj_Last = Output[2]
          I_Last = Output[3]
          Addition <- c(Pi_Last, Pj_Last, I_Last)
          df <- rbind(df, Addition)
          colnames(df) <- c(paste0("Pi_", SimNum[l]),
                            paste0("Pj_", SimNum[l]),
                            paste0("I_", SimNum[l]))
        }
        Inits <- data.frame(1, 0, 1)
        if (IntroPt[j] == 0){
          Inits <- data.frame(1, 0, 1)
        }
        colnames(Inits) <- c(paste0("Pi_", SimNum[l]),
                             paste0("Pj_", SimNum[l]),
                             paste0("I_", SimNum[l]))
        df <- data.frame(rbind(Inits, df))
        SD_df <- data.frame(cbind(SD_df, df))
      }
      List1[[k]] <- SD_df
    }
    List1Names <- c()
    for (o in 1:length(sd)){
      List1Names[o] <- paste0("SD_", sd[o])
    }
    names(List1) <- List1Names
    List2[[j]] <- List1
  }
  List2Names <- c()
  for (p in 1:length(IntroPt)){
    List2Names[p] <- paste0("Intro_", IntroPt[p])
  }
  names(List2) <- List2Names
  List3[[i]] <- List2
}
List3Names <- c()
for (i in 1:length(freq)){
  List3Names[i] <- paste0("Each_", freq[i])
}
names(List3) <- List3Names

saveRDS(List3, file = "Generated_Outputs/DE_HT_PjAd_1P.rds")
