==> 7_Transmission_TwoParasite_AUC.R <==
#### DOC STRING ===============================================================
# This script uses time series outputs to calculate total transmission per
# parasite, for each simulation (1000) of each permutation of the demographically
# and environmentally stochastic model. 

# To run this script, you'll need to (1) re-generate time series data using the 
# scripts in the "Data_Generation" folder and find and replace 
# "EnvironmentalVariability_Repo/Data/EnvStoch-DemEnvStoch_PV_Outputs" with 
# "Generated_Outputs", or (2) download time series data from the FRDR link 
# specified in this repository's README, into 
# "EnvironmentalVariability_Repo/Data/EnvStoch-DemEnvStoch_PV_Outputs". 
# The files are named as follows: DE_LT_Iden.rds, DE_LT_Semi.rds, DE_LT_PiAd.rds, 
# DE_LT_PjAd.rds, DE_HT_Iden.rds, DE_HT_Semi.rds, DE_HT_PiAd.rds, DE_HT_PjAd.rds. 
# Any Outputs will be stored in "EnvironmentalVariability_Repo/Generated_Outputs".

# To see the outputs generated by this script without running it, go to 
# "EnvironmentalVariability_Repo/Outputs/TotalTransmission_Outputs".
#### END DOC STRING ===========================================================

# Required packages
install.packages("dplyr"); install.packages("Bolstad2")
library(dplyr); library(Bolstad2)

TT <- seq(0, 50, 0.1)

#### Low-average temperature environment ####
#### ... Identical TPCs ####
## Raw data file
DE_LT_Iden <- readRDS("Data/EnvStoch-DemEnvStoch_PV_Outputs/DE_LT_Iden.rds")

SimNum <- seq(1, 1000, 1)
Freq <- c(1, 2, 5, seq(10, 100, 10))
IntroPt <- seq(0, 49, 1)
SD <- seq(2, 30, 2)

## Pulling out the permutations we want
DE_LT_Iden_2P <- list()
k = 1
for (i in 1:length(DE_LT_Iden)){
  for (j in 1:length(DE_LT_Iden[[1]][[1]])){
    DE_LT_Iden_2P[[k]] <- DE_LT_Iden[[i]][[1]][[j]]
    k = k + 1
  }
}

## Naming the list elements
Names <- c()
k = 1
for (i in 1:length(DE_LT_Iden)){
  for (j in 1:length(DE_LT_Iden[[1]][[1]])){
    Names[k] <- paste("Mag", SD[j], "_Freq", Freq[i], "_In0", sep = "")
    k = k + 1
  }
}
names(DE_LT_Iden_2P) <- Names

## Calculating the AUC of each simulation, in each list element
Pi_Cols <- seq(2, ncol(DE_LT_Iden_2P[[1]]), 3); Pj_Cols <- seq(3, ncol(DE_LT_Iden_2P[[1]]), 3)
DE_LT_Iden_2P_AUC <- data.frame(Sim = SimNum)
for (i in 1:length(DE_LT_Iden_2P)){
  Vec_Pi <- c(); Vec_Pj <- c()
  for (j in 1:length(Pi_Cols)){
    Pi <- DE_LT_Iden_2P[[i]][ ,Pi_Cols[j]]; Pj <- DE_LT_Iden_2P[[i]][ ,Pj_Cols[j]]
    # Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline"); Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
    Vec_Pi[j] <- sintegral(TT, Pi)$int; Vec_Pj[j] <- sintegral(TT, Pj)$int
  }
  DE_LT_Iden_2P_AUC <- cbind(DE_LT_Iden_2P_AUC, Vec_Pi, Vec_Pj)
}

## Naming the columns
Pi_Cols <- seq(2, ncol(DE_LT_Iden_2P_AUC), 2); Pj_Cols <- seq(3, ncol(DE_LT_Iden_2P_AUC), 2)
Names2 <- c("Sim")
Iters <- rep(seq(1, length(Names), 1), each = 2)
for (i in 2:ncol(DE_LT_Iden_2P_AUC)){
  if (i %in% Pi_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pi", sep = "")}
  if (i %in% Pj_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pj", sep = "")}
}; colnames(DE_LT_Iden_2P_AUC) <- Names2

saveRDS(DE_LT_Iden_2P_AUC, "Generated_Outputs/DE_LT_Iden_2P_AUC.rds")

#### ... Semi-Identical ####
## Raw data file
DE_LT_Semi <- readRDS("Data/EnvStoch-DemEnvStoch_PV_Outputs/DE_LT_Semi.rds")

## Pulling out the permutations we want
DE_LT_Semi_2P <- list()
k = 1
for (i in 1:length(DE_LT_Semi)){
  for (j in 1:length(DE_LT_Semi[[1]][[1]])){
    DE_LT_Semi_2P[[k]] <- DE_LT_Semi[[i]][[1]][[j]]
    k = k + 1
  }
}

## Naming the list elements
Names <- c()
k = 1
for (i in 1:length(DE_LT_Semi)){
  for (j in 1:length(DE_LT_Semi[[1]][[1]])){
    Names[k] <- paste("Mag", SD[j], "_Freq", Freq[i], "_In0", sep = "")
    k = k + 1
  }
}
names(DE_LT_Semi_2P) <- Names

## Calculating the AUC of each simulation, in each list element
Pi_Cols <- seq(2, ncol(DE_LT_Semi_2P[[1]]), 3); Pj_Cols <- seq(3, ncol(DE_LT_Semi_2P[[1]]), 3)
DE_LT_Semi_2P_AUC <- data.frame(Sim = SimNum)
for (i in 1:length(DE_LT_Semi_2P)){
  Vec_Pi <- c(); Vec_Pj <- c()
  for (j in 1:length(Pi_Cols)){
    Pi <- DE_LT_Semi_2P[[i]][ ,Pi_Cols[j]]; Pj <- DE_LT_Semi_2P[[i]][ ,Pj_Cols[j]]
    # Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline"); Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
    Vec_Pi[j] <- sintegral(TT, Pi)$int; Vec_Pj[j] <- sintegral(TT, Pj)$int
  }
  DE_LT_Semi_2P_AUC <- cbind(DE_LT_Semi_2P_AUC, Vec_Pi, Vec_Pj)
}

## Naming the columns
Pi_Cols <- seq(2, ncol(DE_LT_Semi_2P_AUC), 2); Pj_Cols <- seq(3, ncol(DE_LT_Semi_2P_AUC), 2)
Names2 <- c("Sim")
Iters <- rep(seq(1, length(Names), 1), each = 2)
for (i in 2:ncol(DE_LT_Semi_2P_AUC)){
  if (i %in% Pi_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pi", sep = "")}
  if (i %in% Pj_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pj", sep = "")}
}; colnames(DE_LT_Semi_2P_AUC) <- Names2

saveRDS(DE_LT_Semi_2P_AUC, "Generated_Outputs/DE_LT_Semi_2P_AUC.rds")

#### ... Pi-Advantage TPCs ####
## Raw data file
DE_LT_PiAd <- readRDS("Data/EnvStoch-DemEnvStoch_PV_Outputs/DE_LT_PiAd.rds")

## Pulling out the permutations we want
DE_LT_PiAd_2P <- list()
k = 1
for (i in 1:length(DE_LT_PiAd)){
  for (j in 1:length(DE_LT_PiAd[[1]][[1]])){
    DE_LT_PiAd_2P[[k]] <- DE_LT_PiAd[[i]][[1]][[j]]
    k = k + 1
  }
}

## Naming the list elements
Names <- c()
k = 1
for (i in 1:length(DE_LT_PiAd)){
  for (j in 1:length(DE_LT_PiAd[[1]][[1]])){
    Names[k] <- paste("Mag", SD[j], "_Freq", Freq[i], "_In0", sep = "")
    k = k + 1
  }
}
names(DE_LT_PiAd_2P) <- Names

## Calculating the AUC of each simulation, in each list element
Pi_Cols <- seq(2, ncol(DE_LT_PiAd_2P[[1]]), 3); Pj_Cols <- seq(3, ncol(DE_LT_PiAd_2P[[1]]), 3)
DE_LT_PiAd_2P_AUC <- data.frame(Sim = SimNum)
for (i in 1:length(DE_LT_PiAd_2P)){
  Vec_Pi <- c(); Vec_Pj <- c()
  for (j in 1:length(Pi_Cols)){
    Pi <- DE_LT_PiAd_2P[[i]][ ,Pi_Cols[j]]; Pj <- DE_LT_PiAd_2P[[i]][ ,Pj_Cols[j]]
    # Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline"); Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
    Vec_Pi[j] <- sintegral(TT, Pi)$int; Vec_Pj[j] <- sintegral(TT, Pj)$int
  }
  DE_LT_PiAd_2P_AUC <- cbind(DE_LT_PiAd_2P_AUC, Vec_Pi, Vec_Pj)
}

## Naming the columns
Pi_Cols <- seq(2, ncol(DE_LT_PiAd_2P_AUC), 2); Pj_Cols <- seq(3, ncol(DE_LT_PiAd_2P_AUC), 2)
Names2 <- c("Sim")
Iters <- rep(seq(1, length(Names), 1), each = 2)
for (i in 2:ncol(DE_LT_PiAd_2P_AUC)){
  if (i %in% Pi_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pi", sep = "")}
  if (i %in% Pj_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pj", sep = "")}
}; colnames(DE_LT_PiAd_2P_AUC) <- Names2

saveRDS(DE_LT_PiAd_2P_AUC, "Generated_Outputs/DE_LT_PiAd_2P_AUC.rds")

#### ... Pj-Advantage, Version A TPCs ####
## Raw data file
DE_LT_PjAd <- readRDS("Data/EnvStoch-DemEnvStoch_PV_Outputs/DE_LT_PjAd.rds")

## Pulling out the permutations we want
DE_LT_PjAd_2P <- list()
k = 1
for (i in 1:length(DE_LT_PjAd)){
  for (j in 1:length(DE_LT_PjAd[[1]][[1]])){
    DE_LT_PjAd_2P[[k]] <- DE_LT_PjAd[[i]][[1]][[j]]
    k = k + 1
  }
}

## Naming the list elements
Names <- c()
k = 1
for (i in 1:length(DE_LT_PjAd)){
  for (j in 1:length(DE_LT_PjAd[[1]][[1]])){
    Names[k] <- paste("Mag", SD[j], "_Freq", Freq[i], "_In0", sep = "")
    k = k + 1
  }
}
names(DE_LT_PjAd_2P) <- Names

## Calculating the AUC of each simulation, in each list element
Pi_Cols <- seq(2, ncol(DE_LT_PjAd_2P[[1]]), 3); Pj_Cols <- seq(3, ncol(DE_LT_PjAd_2P[[1]]), 3)
DE_LT_PjAd_2P_AUC <- data.frame(Sim = SimNum)
for (i in 1:length(DE_LT_PjAd_2P)){
  Vec_Pi <- c(); Vec_Pj <- c()
  for (j in 1:length(Pi_Cols)){
    Pi <- DE_LT_PjAd_2P[[i]][ ,Pi_Cols[j]]; Pj <- DE_LT_PjAd_2P[[i]][ ,Pj_Cols[j]]
    # Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline"); Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
    Vec_Pi[j] <- sintegral(TT, Pi)$int; Vec_Pj[j] <- sintegral(TT, Pj)$int
  }
  DE_LT_PjAd_2P_AUC <- cbind(DE_LT_PjAd_2P_AUC, Vec_Pi, Vec_Pj)
}

## Naming the columns
Pi_Cols <- seq(2, ncol(DE_LT_PjAd_2P_AUC), 2); Pj_Cols <- seq(3, ncol(DE_LT_PjAd_2P_AUC), 2)
Names2 <- c("Sim")
Iters <- rep(seq(1, length(Names), 1), each = 2)
for (i in 2:ncol(DE_LT_PjAd_2P_AUC)){
  if (i %in% Pi_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pi", sep = "")}
  if (i %in% Pj_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pj", sep = "")}
}; colnames(DE_LT_PjAd_2P_AUC) <- Names2

saveRDS(DE_LT_PjAd_2P_AUC, "Generated_Outputs/DE_LT_PjAd_2P_AUC.rds")

#### High-average temperature environment ####
#### ... Identical TPCs ####
## Raw data file
DE_HT_Iden <- readRDS("Data/EnvStoch-DemEnvStoch_PV_Outputs/DE_HT_Iden.rds")

SD <- seq(2, 40, 2)

## Pulling out the permutations we want
DE_HT_Iden_2P <- list()
k = 1
for (i in 1:length(DE_HT_Iden)){
  for (j in 1:length(DE_HT_Iden[[1]][[1]])){
    DE_HT_Iden_2P[[k]] <- DE_HT_Iden[[i]][[1]][[j]]
    k = k + 1
  }
}

## Naming the list elements
Names <- c()
k = 1
for (i in 1:length(DE_HT_Iden)){
  for (j in 1:length(DE_HT_Iden[[1]][[1]])){
    Names[k] <- paste("Mag", SD[j], "_Freq", Freq[i], "_In0", sep = "")
    k = k + 1
  }
}
names(DE_HT_Iden_2P) <- Names

## Calculating the AUC of each simulation, in each list element
Pi_Cols <- seq(2, ncol(DE_HT_Iden_2P[[1]]), 3); Pj_Cols <- seq(3, ncol(DE_HT_Iden_2P[[1]]), 3)
DE_HT_Iden_2P_AUC <- data.frame(Sim = SimNum)
for (i in 1:length(DE_HT_Iden_2P)){
  Vec_Pi <- c(); Vec_Pj <- c()
  for (j in 1:length(Pi_Cols)){
    Pi <- DE_HT_Iden_2P[[i]][ ,Pi_Cols[j]]; Pj <- DE_HT_Iden_2P[[i]][ ,Pj_Cols[j]]
    # Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline"); Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
    Vec_Pi[j] <- sintegral(TT, Pi)$int; Vec_Pj[j] <- sintegral(TT, Pj)$int
  }
  DE_HT_Iden_2P_AUC <- cbind(DE_HT_Iden_2P_AUC, Vec_Pi, Vec_Pj)
}

## Naming the columns
Pi_Cols <- seq(2, ncol(DE_HT_Iden_2P_AUC), 2); Pj_Cols <- seq(3, ncol(DE_HT_Iden_2P_AUC), 2)
Names2 <- c("Sim")
Iters <- rep(seq(1, length(Names), 1), each = 2)
for (i in 2:ncol(DE_HT_Iden_2P_AUC)){
  if (i %in% Pi_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pi", sep = "")}
  if (i %in% Pj_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pj", sep = "")}
}; colnames(DE_HT_Iden_2P_AUC) <- Names2

saveRDS(DE_HT_Iden_2P_AUC, "Generated_Outputs/DE_HT_Iden_2P_AUC.rds")

#### ... Semi-Identical TPCs ####
## Raw data file
DE_HT_Semi <- readRDS("Data/EnvStoch-DemEnvStoch_PV_Outputs/DE_HT_Semi.rds")

## Pulling out the permutations we want
DE_HT_Semi_2P <- list()
k = 1
for (i in 1:length(DE_HT_Semi)){
  for (j in 1:length(DE_HT_Semi[[1]][[1]])){
    DE_HT_Semi_2P[[k]] <- DE_HT_Semi[[i]][[1]][[j]]
    k = k + 1
  }
}

## Naming the list elements
Names <- c()
k = 1
for (i in 1:length(DE_HT_Semi)){
  for (j in 1:length(DE_HT_Semi[[1]][[1]])){
    Names[k] <- paste("Mag", SD[j], "_Freq", Freq[i], "_In0", sep = "")
    k = k + 1
  }
}
names(DE_HT_Semi_2P) <- Names

## Calculating the AUC of each simulation, in each list element
Pi_Cols <- seq(2, ncol(DE_HT_Semi_2P[[1]]), 3); Pj_Cols <- seq(3, ncol(DE_HT_Semi_2P[[1]]), 3)
DE_HT_Semi_2P_AUC <- data.frame(Sim = SimNum)
for (i in 1:length(DE_HT_Semi_2P)){
  Vec_Pi <- c(); Vec_Pj <- c()
  for (j in 1:length(Pi_Cols)){
    Pi <- DE_HT_Semi_2P[[i]][ ,Pi_Cols[j]]; Pj <- DE_HT_Semi_2P[[i]][ ,Pj_Cols[j]]
    # Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline"); Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
    Vec_Pi[j] <- sintegral(TT, Pi)$int; Vec_Pj[j] <- sintegral(TT, Pj)$int
  }
  DE_HT_Semi_2P_AUC <- cbind(DE_HT_Semi_2P_AUC, Vec_Pi, Vec_Pj)
}

## Naming the columns
Pi_Cols <- seq(2, ncol(DE_HT_Semi_2P_AUC), 2); Pj_Cols <- seq(3, ncol(DE_HT_Semi_2P_AUC), 2)
Names2 <- c("Sim")
Iters <- rep(seq(1, length(Names), 1), each = 2)
for (i in 2:ncol(DE_HT_Semi_2P_AUC)){
  if (i %in% Pi_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pi", sep = "")}
  if (i %in% Pj_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pj", sep = "")}
}; colnames(DE_HT_Semi_2P_AUC) <- Names2

saveRDS(DE_HT_Semi_2P_AUC, "Generated_Outputs/DE_HT_Semi_2P_AUC.rds")

#### ... Pi-Advantage TPCs ####
## Raw data file
DE_HT_PiAd <- readRDS("Data/EnvStoch-DemEnvStoch_PV_Outputs/DE_HT_PiAd.rds")

## Pulling out the permutations we want
DE_HT_PiAd_2P <- list()
k = 1
for (i in 1:length(DE_HT_PiAd)){
  for (j in 1:length(DE_HT_PiAd[[1]][[1]])){
    DE_HT_PiAd_2P[[k]] <- DE_HT_PiAd[[i]][[1]][[j]]
    k = k + 1
  }
}

## Naming the list elements
Names <- c()
k = 1
for (i in 1:length(DE_HT_PiAd)){
  for (j in 1:length(DE_HT_PiAd[[1]][[1]])){
    Names[k] <- paste("Mag", SD[j], "_Freq", Freq[i], "_In0", sep = "")
    k = k + 1
  }
}
names(DE_HT_PiAd_2P) <- Names

## Calculating the AUC of each simulation, in each list element
Pi_Cols <- seq(2, ncol(DE_HT_PiAd_2P[[1]]), 3); Pj_Cols <- seq(3, ncol(DE_HT_PiAd_2P[[1]]), 3)
DE_HT_PiAd_2P_AUC <- data.frame(Sim = SimNum)
for (i in 1:length(DE_HT_PiAd_2P)){
  Vec_Pi <- c(); Vec_Pj <- c()
  for (j in 1:length(Pi_Cols)){
    Pi <- DE_HT_PiAd_2P[[i]][ ,Pi_Cols[j]]; Pj <- DE_HT_PiAd_2P[[i]][ ,Pj_Cols[j]]
    # Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline"); Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
    Vec_Pi[j] <- sintegral(TT, Pi)$int; Vec_Pj[j] <- sintegral(TT, Pj)$int
  }
  DE_HT_PiAd_2P_AUC <- cbind(DE_HT_PiAd_2P_AUC, Vec_Pi, Vec_Pj)
}

## Naming the columns
Pi_Cols <- seq(2, ncol(DE_HT_PiAd_2P_AUC), 2); Pj_Cols <- seq(3, ncol(DE_HT_PiAd_2P_AUC), 2)
Names2 <- c("Sim")
Iters <- rep(seq(1, length(Names), 1), each = 2)
for (i in 2:ncol(DE_HT_PiAd_2P_AUC)){
  if (i %in% Pi_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pi", sep = "")}
  if (i %in% Pj_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pj", sep = "")}
}; colnames(DE_HT_PiAd_2P_AUC) <- Names2

saveRDS(DE_HT_PiAd_2P_AUC, "Generated_Outputs/DE_HT_PiAd_2P_AUC.rds")

#### ... Pj-Advantage, Version A TPCs ####
## Raw data file
DE_HT_PjAd <- readRDS("Data/EnvStoch-DemEnvStoch_PV_Outputs/DE_HT_PjAd.rds")

## Pulling out the permutations we want
DE_HT_PjAd_2P <- list()
k = 1
for (i in 1:length(DE_HT_PjAd)){
  for (j in 1:length(DE_HT_PjAd[[1]][[1]])){
    DE_HT_PjAd_2P[[k]] <- DE_HT_PjAd[[i]][[1]][[j]]
    k = k + 1
  }
}

## Naming the list elements
Names <- c()
k = 1
for (i in 1:length(DE_HT_PjAd)){
  for (j in 1:length(DE_HT_PjAd[[1]][[1]])){
    Names[k] <- paste("Mag", SD[j], "_Freq", Freq[i], "_In0", sep = "")
    k = k + 1
  }
}
names(DE_HT_PjAd_2P) <- Names

## Calculating the AUC of each simulation, in each list element
Pi_Cols <- seq(2, ncol(DE_HT_PjAd_2P[[1]]), 3); Pj_Cols <- seq(3, ncol(DE_HT_PjAd_2P[[1]]), 3)
DE_HT_PjAd_2P_AUC <- data.frame(Sim = SimNum)
for (i in 1:length(DE_HT_PjAd_2P)){
  Vec_Pi <- c(); Vec_Pj <- c()
  for (j in 1:length(Pi_Cols)){
    Pi <- DE_HT_PjAd_2P[[i]][ ,Pi_Cols[j]]; Pj <- DE_HT_PjAd_2P[[i]][ ,Pj_Cols[j]]
    # Vec_Pi[j] <- auc(x = TT, y = Pi, type = "spline"); Vec_Pj[j] <- auc(x = TT, y = Pj, type = "spline")
    Vec_Pi[j] <- sintegral(TT, Pi)$int; Vec_Pj[j] <- sintegral(TT, Pj)$int
  }
  DE_HT_PjAd_2P_AUC <- cbind(DE_HT_PjAd_2P_AUC, Vec_Pi, Vec_Pj)
}

## Naming the columns
Pi_Cols <- seq(2, ncol(DE_HT_PjAd_2P_AUC), 2); Pj_Cols <- seq(3, ncol(DE_HT_PjAd_2P_AUC), 2)
Names2 <- c("Sim")
Iters <- rep(seq(1, length(Names), 1), each = 2)
for (i in 2:ncol(DE_HT_PjAd_2P_AUC)){
  if (i %in% Pi_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pi", sep = "")}
  if (i %in% Pj_Cols){Names2[i] <- paste(Names[Iters[i-1]], "_Pj", sep = "")}
}; colnames(DE_HT_PjAd_2P_AUC) <- Names2

saveRDS(DE_HT_PjAd_2P_AUC, "Generated_Outputs/DE_HT_PjAd_2P_AUC.rds")
